requirements.txt
absl-py==2.2.2
addict==2.4.0
antlr4-python3-runtime==4.8
basicsr==1.4.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
cmake==4.0.0
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
einops==0.8.1
filelock==3.16.1
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.0
future==1.0.0
google-auth==2.40.1
google-auth-oauthlib==1.0.0
grpcio==1.70.0
humanfriendly==10.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
kornia==0.6.12
lit==18.1.8
lmdb==1.6.2
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.4.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.2
onnx==1.17.0
onnxruntime-gpu==1.19.2
opencv-python==4.11.0.86
opencv-python-headless==4.5.5.64
packaging==25.0
pandas==1.3.5
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-image==0.19.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.17.0
sympy==1.13.3
tb-nightly==2.14.0a20230808
tensorboard-data-server==0.7.2
tensorrt==10.1.0
tensorrt-cu12-bindings==10.1.0
tensorrt-cu12-libs==10.1.0
tensorrt_cu12==10.8.0.43
threadpoolctl==3.5.0
tifffile==2023.7.10
tomli==2.2.1
torch==2.4.1
torch_tensorrt==2.4.0
torchfile==0.1.0
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.67.1
triton==3.0.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
Werkzeug==3.0.6
yapf==0.43.0
zipp==3.20.2
